add: carousel components & apply review
