+++ whenthen-web/src/components/CarouselComponent/CarouselItem/DesktopDefault.module.scss
@@ -1,0 +1,10 @@
+.carouselItem {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  height: 200px;
+  width: 100%;
+  position: absoulte;
+  background-color: white;
+  color: #fff;
+}
