+++ whenthen-web/src/components/CarouselComponent/DesktopDefault.module.scss
@@ -8,14 +8,3 @@
   transition: transform 0.1s ease;
   will-change: transform;
 }
-
-.carouselItem {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  height: 200px;
-  width: 100%;
-  position: absoulte;
-  background-color: white;
-  color: #fff;
-}
