+++ whenthen-web/src/components/CarouselComponent/index.tsx
@@ -1,5 +1,10 @@
-import React, { useState, useEffect, useRef } from 'react';
+import { useState } from 'react';
+
+import CarouselItemComponent from './CarouselItem/index';
+import useInterval from '../../hooks/useInterval';
+
 import stylesDesktopDefault from './DesktopDefault.module.scss';
+import useRootData from '../../hooks/useRootData';
 
 const imgSrc = [
   'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQLpZ3AmUlI_h9TNBB-lEnxyFDIplBHlYyL1A&usqp=CAU',
@@ -7,49 +12,24 @@ const imgSrc = [
   'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQ4zRpZR__dX85rRE56oziBQHW9SmLNHK4AOw&usqp=CAU',
 ];
 
-const useInterval = (callback: Function, delay: number) => {
-  const savedCallback = useRef<Function>(callback);
-
-  useEffect(() => {
-    savedCallback.current = callback;
-  }, [callback]);
-
-  useEffect(() => {
-    function tick() {
-      savedCallback.current();
-    }
-    if (delay !== null) {
-      let id = setInterval(tick, delay);
-      return () => clearInterval(id);
-    }
-  }, [delay]);
-};
-
-export interface ProbsCarouselItemComponent {
-  src: string;
-}
-
-const CarouselItemComponent: React.FunctionComponent<
-  ProbsCarouselItemComponent
-> = ({ src }) => {
-  return (
-    <div className={stylesDesktopDefault.carouselItem}>
-      <img src={src} style={{ width: '100%', height: '100%' }} />
-    </div>
-  );
-};
-
 const CarouselComponent = () => {
   const [activeIndex, setActiveIndex] = useState(0);
+
+  const { screenClass } = useRootData(({ appStore }) => ({
+    screenClass: appStore.screenClass.get(),
+  }));
+  const isDesktop = screenClass === 'xl';
+
+  const styles = isDesktop ? stylesDesktopDefault : stylesDesktopDefault;
 
   useInterval(() => {
     setActiveIndex((activeIndex + 1) % imgSrc.length);
   }, 2000);
 
   return (
-    <div className={stylesDesktopDefault.carousel}>
+    <div className={styles.carousel}>
       <div
-        className={stylesDesktopDefault.carouselInner}
+        className={styles.carouselInner}
         style={{ transform: `translateX(-${activeIndex * 100}%)` }}
       >
         {imgSrc.map((src) => {
