+++ whenthen-web/src/hooks/useInterval.ts
@@ -1,0 +1,21 @@
+import { useEffect, useRef } from 'react';
+
+const useInterval = (callback: Function, delay: number) => {
+  const savedCallback = useRef<Function>(callback);
+
+  useEffect(() => {
+    savedCallback.current = callback;
+  }, [callback]);
+
+  useEffect(() => {
+    function tick() {
+      savedCallback.current();
+    }
+    if (delay !== null) {
+      let id = setInterval(tick, delay);
+      return () => clearInterval(id);
+    }
+  }, [delay]);
+};
+
+export default useInterval;
